% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{births1dt}
\alias{births1dt}
\alias{birthsproj1dt}
\title{Annual Birth Counts in Thousands}
\format{
Data table in long format with variables:
\describe{
  \item{country_code}{int Numerical Location Code (3-digit codes following ISO 3166-1 numeric standard)} 
  \item{name}{chr Name of country or region}
  \item{year}{int Year when births occured}
  \item{births}{dbl Birth counts in thousands}
  \item{births_80l, births_95l}{dbl Lower bounds of the 80 and 95 percent probability intervals}
  \item{births_80u, births_95u}{dbl Upper bounds of the 80 and 95 percent probability intervals}
  
}

An object of class \code{data.table} (inherits from \code{data.frame}) with 22946 rows and 8 columns.
}
\source{
These datasets are based on estimates and projections of United Nations, 
    Department of Economic and Social Affairs, Population Division (2024).
}
\usage{
births1dt

birthsproj1dt
}
\description{
Datasets containing the United Nations' annual time series of the birth counts 
for all countries as available in 2024.
}
\details{
\code{births1dt} contains estimated births while \code{birthsproj1dt}
    contains projected births for future years and include 80\% and 95\% 
    probability intervals.
}
\references{
World Population Prospects: The 2024 Revision. \url{http://population.un.org/wpp}.
}
\keyword{datasets}
