% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbr1dt}
\alias{cbr1dt}
\alias{cbrproj1dt}
\title{Annual Crude Birth Rate}
\format{
Data table in long format with variables:
\describe{
  \item{country_code}{int Numerical Location Code (3-digit codes following ISO 3166-1 numeric standard)} 
  \item{name}{chr Name of country or region}
  \item{year}{int Year when CBR occured}
  \item{cbr}{dbl CBR counts in thousands}
  \item{cbr_80l, cbr_95l}{dbl Lower bounds of the 80 and 95 percent probability intervals}
  \item{cbr_80u, cbr_95u}{dbl Upper bounds of the 80 and 95 percent probability intervals}
  
}

An object of class \code{data.table} (inherits from \code{data.frame}) with 22946 rows and 8 columns.
}
\source{
These datasets are based on estimates and projections of United Nations, 
    Department of Economic and Social Affairs, Population Division (2024).
}
\usage{
cbr1dt

cbrproj1dt
}
\description{
Datasets containing the United Nations' annual time series of the crude birth rate (CBR)
for all countries as available in 2024.
}
\details{
\code{cbr1dt} contains estimated CBR while \code{cbrproj1dt}
    contains projected CBR for future years and include 80\% and 95\% 
    probability intervals.
}
\references{
World Population Prospects: The 2024 Revision. \url{http://population.un.org/wpp}.
}
\keyword{datasets}
