% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdr1dt}
\alias{cdr1dt}
\alias{cdrproj1dt}
\title{Annual Crude Death Rate}
\format{
Data table in long format with variables:
\describe{
  \item{country_code}{int Numerical Location Code (3-digit codes following ISO 3166-1 numeric standard)} 
  \item{name}{chr Name of country or region}
  \item{year}{int Year when CDR occured}
  \item{cdr}{dbl CDR counts in thousands}
  \item{cdr_80l, cdr_95l}{dbl Lower bounds of the 80 and 95 percent probability intervals}
  \item{cdr_80u, cdr_95u}{dbl Upper bounds of the 80 and 95 percent probability intervals}
}

An object of class \code{data.table} (inherits from \code{data.frame}) with 22946 rows and 8 columns.
}
\source{
These datasets are based on estimates and projections of United Nations, 
    Department of Economic and Social Affairs, Population Division (2024).
}
\usage{
cdr1dt

cdrproj1dt
}
\description{
Datasets containing the United Nations' annual time series of the crude death rate (CDR)
for all countries as available in 2024.
}
\details{
\code{cdr1dt} contains estimated CDR while \code{cdrproj1dt}
    contains projected CDR for future years and include 80\% and 95\% 
    probability intervals.
}
\references{
World Population Prospects: The 2024 Revision. \url{http://population.un.org/wpp}.
}
\keyword{datasets}
