% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oadrproj1dt}
\alias{oadrproj1dt}
\title{Annual Old Age Dependency Ratio}
\format{
Data table in long format with variables:
\describe{
  \item{country_code}{int Numerical Location Code (3-digit codes following ISO 3166-1 numeric standard)} 
  \item{name}{chr Name of country or region}
  \item{year}{int Year when OADR occured}
  \item{age}{chr Gives the definition of OADR. There are six definitions: 60+/15-59, 65+/15-64, 65+/20-64,
          65+/25-64, 70+/20-69, 70+/25-69.}
  \item{oadr, oadrM, oadrF}{dbl Median OADR computed over the total, male and female, respectively.}
  \item{oadr_80l, oadr_95l, oadrM_80l, oadrM_95l, oadrF_80l, oadrF_95l}{dbl Lower bounds
         of the 80 and 95 percent probability intervals for the total, male and female, respectively.}
 \item{oadr_80u, oadr_95u, oadrM_80u, oadrM_95u, oadrF_80u, oadrF_95u}{dbl Upper bounds
         of the 80 and 95 percent probability intervals for the total, male and female, respectively.}
}
}
\source{
These datasets are based on estimates and projections of United Nations, 
    Department of Economic and Social Affairs, Population Division (2024).
}
\usage{
oadrproj1dt
}
\description{
Dataset containing the United Nations' projected annual time series of the old age dependency ratios (OADR)
for all countries as available in 2024, using multiple definitions.
}
\references{
World Population Prospects: The 2024 Revision. \url{http://population.un.org/wpp}.
}
\keyword{datasets}
