% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{popprojAgeGrp1dt}
\alias{popprojAgeGrp1dt}
\title{Projection of Population Counts by Broader Age Groups}
\format{
Data table in long format with variables:
\describe{
  \item{country_code}{int Numerical Location Code (3-digit codes following ISO 3166-1 numeric standard)} 
  \item{name}{chr Name of country or region}
  \item{year}{int Year}
  \item{age}{chr Age interval}
  \item{pop, popM, popF}{dbl Median population count in thousand for the total, male and female, respectively.}
  \item{pop_80l, pop_95l, popM_80l, popM_95l, popF_80l, popF_95l}{dbl Lower bounds of the 80 and 95 percent probability intervals
         for the total, male and female, respectively.}
  \item{pop_80u, pop_95u, popM_80u, popM_95u, popF_80u, popF_95u}{dbl Upper bounds of the 80 and 95 percent probability intervals
         for the total, male and female, respectively.}
}
}
\source{
These datasets are based on estimates and projections of United Nations, 
    Department of Economic and Social Affairs, Population Division (2024).
}
\usage{
popprojAgeGrp1dt
}
\description{
Annual time series of projected population counts by broader age groups
for all countries, including probability intervals, as available in 2024. These are considered as mid-year counts.
}
\references{
World Population Prospects: The 2024 Revision. \url{http://population.un.org/wpp}.
}
\keyword{datasets}
