% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp2024extra.R
\docType{package}
\name{wpp2024extra-package}
\alias{wpp2024extra-package}
\alias{wpp2024extra}
\title{Additional Datasets from the World Population Prospects 2024}
\description{
Provides additional data to the wpp2024 package from the United Nation's World Population Prospects 2024.
}
