% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yadrproj1dt}
\alias{yadrproj1dt}
\title{Annual Young Age Dependency Ratio}
\format{
Data table in long format with variables:
\describe{
  \item{country_code}{int Numerical Location Code (3-digit codes following ISO 3166-1 numeric standard)} 
  \item{name}{chr Name of country or region}
  \item{year}{int Year when YADR occured}
  \item{age}{chr Gives the definition of YADR. There are six definitions: 0-14/15-59,  0-14/15-64,
      0-19/20-64, 0-19/20-69, 0-24/25-64, 0-24/25-69.}
  \item{yadr, yadrM, yadrF}{dbl Median YADR computed over the total, male and female, respectively.}
  \item{yadr_80l, yadr_95l, yadrM_80l, yadrM_95l, yadrF_80l, yadrF_95l}{dbl Lower bounds
         of the 80 and 95 percent probability intervals for the total, male and female, respectively.}
 \item{yadr_80u, yadr_95u, yadrM_80u, yadrM_95u, yadrF_80u, yadrF_95u}{dbl Upper bounds
         of the 80 and 95 percent probability intervals for the total, male and female, respectively.}
}
}
\source{
These datasets are based on estimates and projections of United Nations, 
    Department of Economic and Social Affairs, Population Division (2024).
}
\usage{
yadrproj1dt
}
\description{
Dataset containing the United Nations' projected annual time series of the young age dependency ratios (YADR)
for all countries as available in 2024, using multiple definitions.
}
\references{
World Population Prospects: The 2024 Revision. \url{http://population.un.org/wpp}.
}
\keyword{datasets}
